인덱스 수정, memberRole 수정
인덱스 int -> Long으로 수정
Member 내 roleSet에 @Enumerated(EnumType.STRING) 옵션 추가
service, repository id Integer -> Long으로 변경

+++ src/main/java/com/ezen/management/controller/CounselingController.java
@@ -111,7 +111,7 @@ public class CounselingController {
 
     //삭제하기
     @PostMapping("/delete")
-    public void delete(int idx){
+    public void delete(Long idx){
 
         log.info("counseling delete gogo");
